Migrate to modular input approach for sections
Previously all of the sections for my resume were located all in one
file.

This is not very modular and would lead to an unmanageable document over
time. This commit moves each section to their own tex file which leads
to a main "driver" resume.tex file, and then the main content is
contained in their respective tex files.